Add search form

+++ db/seeds.rb
@@ -6,10 +6,12 @@
 #   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
 #   Character.create(name: 'Luke', movie: movies.first)
 
-Airport.delete_all
-Flight.delete_all
+if Rails.env.development?
+  Airport.delete_all
+  Flight.delete_all
 
-airport1 = Airport.create(airport_code: 'SFO')
-airport2 = Airport.create(airport_code: 'NYC')
+  airport1 = Airport.create(airport_code: 'SFO', city: 'San Francisco')
+  airport2 = Airport.create(airport_code: 'NYC', city: 'New York')
 
-flight = Flight.create(from_id: airport1.id, to_id: airport2.id, datetime: DateTime.new(DateTime.current.year, DateTime.current.month, DateTime.current.day).change(day: DateTime.current.day + 14))
+  flight = Flight.create(from_id: airport1.id, to_id: airport2.id, datetime: DateTime.new(DateTime.current.year, DateTime.current.month, DateTime.current.day).change(day: DateTime.current.day + 14))
+end
